Create class for processing and clustering images
An application was created that allows for the clustering of images using
face recognition, using facenet's pre-trained model. Most of the code
comes from the 'batchsnap_sorter' repository, but it has been cleaned up
properly. New functions have also been written for pre-processing the images
so that blurriness and lighting conditions do not affect the encoding process
as much.

A custom distance metric is also being used, taken from the following link -
https://gist.github.com/StuartGordonReid/90c8301f9403707a22ff
Using the fractional_distance metric with a power of 4 seems to give optimal
results. Mahalanobis distance was tried, but it didn't make much of a
difference. Also, the clustering doesn't seem to depend much upon whether the
encodings are normalized or not. However, the normalized data will be used in
the future for clustering, in case it is required later for other purposes.