Create application for finding clusters corresponding to a given person
A new class has been created, allowing to capture images of a person and
therefore search for that person among the corresponding clusters. This is
done by computing the mean encoding of a face taken from a webcam (over a
series of frames), and then checking whether this lies within the 512D
hypersphere formed by the precomputed mean encodings of each cluster. The
radius is generated by the points that lie within 1.5 standard deviations of
the mean encoding for that cluster.

This therefore gives a list of possible clusters, narrowing down the number
of clusters from potentially hundreds to a few.