Create class for clustering videos
Created a new class that allows for processing and clustering faces found in
videos. Every alternate frame is processed to cut processing time in half.
It may be possible to parallelize this operation to increase speed. Currently,
it is not integrated with the ImageClusterer class - this integration will be
added in the future.